+++ Actividad4_5/src/main/java/org/zabalburu/daw1/actividad4_5/modelo/Alumno.java
@@ -45,9 +45,9 @@ public class Alumno {
                 sumaNotas += notaAlumnoProcesado;
             }
         }
-        
-        double notaMedia = sumaNotas / contadorAlumnos;
-        
+
+        double notaMedia = sumaNotas / (double) contadorAlumnos;
+
         JOptionPane.showMessageDialog(null, """
                                             <html>
                                                 <hr />
